requirements.txt
asgiref==3.8.1
cryptography==43.0.1
Django==5.1.2
django-environ==0.11.2
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-secret-manager==2.21.0
google-cloud-storage==2.18.2
google-cloud-texttospeech==2.21.1
gunicorn==23.0.0
packaging==24.1
PyJWT==2.9.0
pylint==3.3.1
requests==2.32.3
sqlparse==0.5.1